Fix type hints

+++ hc/front/templatetags/hc_extras.py
@@ -185,7 +185,7 @@ def now_isoformat() -> str:
 
 
 @register.filter
-def timestamp(dt: datetime) -> int:
+def timestamp(dt: datetime) -> float:
     return dt.timestamp()
